toddy_shop_backend/pyproject.toml
[project]
name = "KeralaToddyFinderBackend"
version = "0.1.0"
description = "Django backend for Kerala Toddy Finder App"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.3.1",
    "django-filter>=24.0",
    "Pillow>=11.0.0",
    "drf-spectacular>=0.29.0",
    "whitenoise>=6.0.0",
    "psycopg[binary]>=3.1.18",
    "django-import-export>=4.4.0"
]

[tool.black]
line-length = 119
exclude = '''
/(
    migrations
  | venv
  | \.venv
)/
'''
